Updating test

+++ rest-api/dao/site_dao.py
@@ -17,7 +17,7 @@ class _FhirSite(FhirMixin, BackboneElement):
   resource_name = 'Site'
   _PROPERTIES = [
     FhirProperty('display_name', str, required=True),
-    FhirProperty('mayolink_client_number', long, required=True),
+    FhirProperty('mayolink_client_number', int, required=True),
     FhirProperty('site_status', str, required=True),
     FhirProperty('launch_date', fhirdate.FHIRDate),
     FhirProperty('notes', str),
